+++ src/Codeception/Lib/InnerBrowser.php
@@ -638,8 +638,8 @@ class InnerBrowser extends Module implements Web
         } catch (ParseException $e) {
         }
         if (!Locator::isXPath($selector)) {
-//            return new \Symfony\Component\DomCrawler\Crawler;
-            throw new TestRuntime("XPath `$selector` is malformed");
+            codecept_debug("XPath `$selector` is malformed!");
+            return new \Symfony\Component\DomCrawler\Crawler;
         }
 
         return @$this->crawler->filterXPath($selector);