+++ CHANGELOG.md
@@ -13,7 +13,7 @@ settings:
 ```
 
 * Fixed retrieval of codecoverage reports on remote server #1379 *2014-10-06*
-* [PhpBrowser][Frameworks] Descriptive exception is thrown on XPath syntax error. Fixes #1409 *2014-10-06*
+* [PhpBrowser][Frameworks] Malformed XPath won't throw fatal error, but makes tests fail. Fixes #1409 *2014-10-06*
 * Build command generates actors for included suites. See #1267 *2014-10-03*
 * CodeCoverage throws error on unsuccessful requests (status code is not 200) to remote server. Fixes #346 *2014-10-03*
 * CodeCoverage can be disabled per suite. Fix #1249 *2014-10-02*